fix code coverage

+++ backend/model/user.js
@@ -1,20 +1,14 @@
-/* eslint-disable no-console */
 const { ObjectId } = require('mongodb');
 const { getDB } = require('./db');
 
-/**
-* Add a new user to the collection
-* @param {*} newUser - new user to add
-* @returns
-*/
 const addUser = async (newUser) => {
+  if (!newUser.username || !newUser.password) {
+    throw new Error('Missing fields');
+  }
   try {
-    // get the db
     const db = await getDB();
     const result = await db.collection('users').insertOne(newUser);
-    // print the id of the user
     console.log(`New user created with id: ${result.insertedId}`);
-    // return the result
     return result.insertedId;
   } catch (err) {
     console.error(err);
@@ -22,16 +16,10 @@ const addUser = async (newUser) => {
   }
 };
 
-/**
-  * get all users in a collection
-  * @returns
-  */
 const getAllUsers = async () => {
   try {
-    // get the db
     const db = await getDB();
     const result = await db.collection('users').find({}).toArray();
-    // print the results
     console.log(`Users: ${JSON.stringify(result)}`);
     return result;
   } catch (err) {
@@ -40,17 +28,13 @@ const getAllUsers = async () => {
   }
 };
 
-/**
-  * GET/READ a user given their ID
-  * @param {*} userID - user id of a user
-  * @returns
-  */
 const getUser = async (userID) => {
+  if (!ObjectId.isValid(userID)) {
+    throw new Error('Invalid user ID');
+  }
   try {
-    // get the db
     const db = await getDB();
     const result = await db.collection('users').findOne({ _id: new ObjectId(userID) });
-    // print the result
     console.log(`User: ${JSON.stringify(result)}`);
     return result;
   } catch (err) {
@@ -59,17 +43,10 @@ const getUser = async (userID) => {
   }
 };
 
-/**
-  * get a user by username
-  * @param {*} username - the username of a user
-  * @returns
-  */
 const getUserByUName = async (username) => {
   try {
-    // get the db
     const db = await getDB();
     const result = await db.collection('users').findOne({ username });
-    // print the result
     console.log(`User: ${JSON.stringify(result)}`);
     return result;
   } catch (err) {
@@ -78,15 +55,11 @@ const getUserByUName = async (username) => {
   }
 };
 
-/**
-* Update a user's username
-* @param {*} userID - the id of a user
-* @param {*} newPassword - new password for a user
-* @returns
-*/
 const updateUser = async (userID, newPassword) => {
+  if (!ObjectId.isValid(userID)) {
+    throw new Error('Invalid user ID');
+  }
   try {
-    // get the db
     const db = await getDB();
     const result = await db.collection('users').updateOne(
       { _id: new ObjectId(userID) },
@@ -99,19 +72,13 @@ const updateUser = async (userID, newPassword) => {
   }
 };
 
-/**
-* Delete a user with userID
-* @param {*} userID - the userID to delete
-* @returns
-*/
 const deleteUser = async (userID) => {
+  if (!ObjectId.isValid(userID)) {
+    throw new Error('Invalid user ID');
+  }
   try {
-    // get the db
     const db = await getDB();
-    const result = await db.collection('users').deleteOne(
-      { _id: new ObjectId(userID) },
-    );
-    // print the result
+    const result = await db.collection('users').deleteOne({ _id: new ObjectId(userID) });
     console.log(`User: ${JSON.stringify(result)}`);
     return result;
   } catch (err) {
@@ -120,7 +87,6 @@ const deleteUser = async (userID) => {
   }
 };
 
-// export the functions
 module.exports = {
   addUser,
   getAllUsers,
